Fix explicitly convert literals and variables to expected types NSInteger and NSUInteger

+++ Source/Definition/Method/TyphoonMethod.m
@@ -92,7 +92,7 @@
     NSUInteger hash = (NSUInteger) sel_getName(_selector);
 
     for (id <TyphoonParameterInjection> parameter in _injectedParameters) {
-        hash = (NSUInteger) ((5 << hash) - hash + [[parameter description] hash]);
+        hash = (NSUInteger) (((NSUInteger)5 << hash) - hash + [[parameter description] hash]);
     }
 
     return hash;
